PR changes:   - out parameter removed from ReadArray(Map)LengthOrNull methods;   - WriteArray(Map)HeaderAndLength method now writes length as uint (msgpack specification);

+++ src/msgpack/Converters/ReadOnlyListConverter.cs
@@ -14,7 +14,7 @@ namespace TarantoolDnx.MsgPack.Converters
                 return;
             }
 
-            writer.WriteArrayHeaderAndLength(value.Count);
+            writer.WriteArrayHeaderAndLength((uint) value.Count);
             var elementConverter = context.GetConverter<TElement>();
             ValidateConverter(elementConverter);
